==> notebooks/2.0-processing_and_cleaning.ipynb <==
---------------------------------------------------------------
## III. PRÉTRAITEMENT DES DONNÉES
--------------------------------------------------------------

1. Nettoyer caractères spéciaux, url, hashtags, ...
2. Dédoublonner avec Sequence Matcher


Principe : (1) Au-delà des doublons parfaits que l'on peut supprimer avec deduplicate(), on a des preque-doublons, qui se distinguent par qqs caractères spéciaux, un commentaire ou @username après retweet. (2) Deux textes ne peuvent pas être des doublons si leurs longueurs sont trop différentes (seuil empirique) (3) au-delà du traitement de 45000 tweets, le processing est soit très long (plusieurs heures), sinon on est limité par la capacité RAM.

* (a) Éliminer doublons parfaits.
* (b) Nettoyer l'ensemble
* (c) Couper en deux parties selon longueur.
* (d) Dédoublonner avec SequenceMatcher
* (e) Concatener
* (f) Extraire une partie au milieu.
* (g) Dédoublonner cette partie.
* (h) Reconcatener
----------------------------------------------------------------

import pandas as pd
import numpy as np
import time

import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
nltk.download('punkt')
nltk.download('stopwords')

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
from difflib import SequenceMatcher

# Import du jeu de données
df_all = pd.read_csv('../data/raw/df_full_source.csv')

# Affichage des informations du jeu de données
df_all.info()

# On supprime les colonnes inutiles et renomme les autres
df_all.drop('Unnamed: 0.1',axis = 1, inplace = True)
df_all = df_all.reset_index()
df_all.columns = df_all.columns.str.replace("Unnamed: 0", "id_origin")
df_all.head(3)

---
### 1. Premier traitement des données en deux parties
---

#### 1.1. Sélectionner les extractions à traiter (longueur max = env. 45 000, limite RAM)

# Éliminer les doublons parfaits
# -------------------------------------
df_all = df_all.drop_duplicates(subset=['text'], keep='first')
# 102661 - 76319 = 26342 doublons bruts
print('df dédoublonné :', len(df_all))
df_all.loc[:,'sign_count'] = df_all['text'].apply(lambda x : len(x))
df_all.head(3)

#### 1.2. Nettoyage du texte d'origine et création d'une colonne de text_prepr(ocessing)

df = df_all.copy()

# Nettoyage : on crée une colonne 'text_prepr' et on laisse le 'text' original tranquille
# ----------------------------------------------------------------------------------------
df['text_prepr'] = df.loc[:,'text'].str.replace('WASHINGTON', ' ')
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace(r'Reuters|reuters|REUTERS', ' ', regex=True)
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace(r'Ä¶|Äô|Äù|Äú|Å©|äî|Äî', ' ', regex=True) #lié à la traduction depuis html. il y a probablement une meilleure méthode.
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace(r'\bu\b', ' ', regex=True) # supprimer ou remplacer les 'u' qui représentent 'you'
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace(r'\bs\b', ' ', regex=True) # supprimer les 's' qui représenent 'is' ou 'has'
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace(r'\br\b', ' ', regex=True) # supprimer ou remplacer les 'r' qui représentent 'are'

# supprimer url, e-mail, special characters
import re
def remove_urls(text):
    if isinstance(text, str):
        url_pattern = re.compile(r'http[s]?://\S+|www\.\S+')
        return url_pattern.sub('', text)

def remove_Emails(text):
    if isinstance(text, str):
        Email_pattern = re.compile(r'([a-zA-Z0-9_\.-]+)@([a-zA-Z0-9_\.-]+)\.([a-zA-Z]{2,5})')
        return Email_pattern.sub('', text)

def remove_gobbledegook(text):
    if isinstance(text, str):
    # Supprimer les strings long composés de chiffres et de lettres (sans espaces, des codes d'identification probablement)
        GobblGook = re.compile(r'\b(?=[a-zA-Z0-9]*[A-Z])(?=[a-zA-Z0-9]*[a-z])(?=[a-zA-Z0-9]*\d)[a-zA-Z0-9]{8,12}\b')
        text = GobblGook.sub(' ', text)
        return text

def remove_ATusername(text):
    if isinstance(text, str):
    # Supprimer les @username Twitter
        ATusername = re.compile(r'@([a-zA-Z0-9_\.-]+)')
        text = ATusername.sub(' ', text)
        return text

def remove_hashtag(text) :
    if isinstance(text, str):
    # Supprimer hastags par ex. #blizzard2016
        ATusername = re.compile(r'#([a-zA-Z0-9_\.-]+)')
        text = ATusername.sub(' ', text)
        return text

def remove_speCar_exclu_comma_dot(text):
    if isinstance(text, str):
    # Supprimer tous les caractères spéciaux et ponctuation sauf points et virgules (en cas de sentence tokenization)
        remove_punctuation = re.compile(r'[^\w\s]')
        text = remove_punctuation.sub(' ', text)
        return text

df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_urls)
df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_Emails)
df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_gobbledegook)
df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_ATusername)
df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_hashtag)
df.loc[:,'text_prepr'] = df['text_prepr'].apply(remove_speCar_exclu_comma_dot)
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace('.', '. ') # parfois certaines phrases successives collent l'une à l'autre.
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace('  ', ' ')
df.loc[:,'text_prepr'] = df['text_prepr'].str.replace('   ', ' ')

#### 1.3. Fonction SequenceMatcher pour calculer similarités fortes entre textes

df = df.sort_values(by='sign_count')
df.head(3)
df_1_2=df[:43000].reset_index(drop=True)
df_2_2=df[43000:].reset_index(drop=True)
print(len(df_1_2))
print(len(df_2_2))
df_1_2.head(3)


def compute_similarity(text1, text2):
    """Calcule la similarité de deux textes en utilisant la distance de Levenshtein."""
    return SequenceMatcher(None, text1, text2).ratio()

def flag_similar_posts(df, threshold=0.9,max_features=15000):
    """Flague les lignes avec une similarité lexicale supérieure au seuil donné."""

    # Initialisation des colonnes nécessaires
    df.loc[:,'has_similarity'] = 0
    df.loc[:,'similar_with'] = None
    df.loc[:,'similarity_group'] = None
    df.loc[:,'similarity_score'] = 0.0

    # Calcul du TF-IDF
    vectorizer = TfidfVectorizer(max_features=max_features).fit_transform(df['text_prepr'])
    vectors = vectorizer.toarray()

    # Calcul de la similarité cosinus
    cosine_sim = cosine_similarity(vectors)

    group_counter = 0
    visited = np.zeros(len(df), dtype=bool)

    for i in range(len(df)):
        if visited[i]:
            continue

        # Identifie les textes similaires
        similar_indices = np.where(cosine_sim[i] >= threshold)[0]
        if len(similar_indices) > 1:
            group_counter += 1

            for j in similar_indices:
                if i != j:
                    df.at[j, 'has_similarity'] = 1
                    df.at[j, 'similar_with'] = df.at[i, 'id_origin']
                    df.at[j, 'similarity_score'] = cosine_sim[i][j]
                    df.at[j, 'similarity_group'] = group_counter
                    visited[j] = True

            # Pour le premier texte lui-même
            df.at[i, 'has_similarity'] = 1
            df.at[i, 'similarity_score'] = 1.0  # Car il est identique à lui-même
            df.at[i, 'similarity_group'] = group_counter

    return df

#### 1.4. Exécuter sequence matcher et identifier les doublons pour les DEUX moitiés

**1 sur 2** : # PREMIERE MOITIÉ


# Seq Matcher 1 sur 2 : threshold = 0.9
# ---------------------------------------
start = time.time()
flag_similar_posts(df_1_2)
df.head(3)
stop = time.time()
print(f"Temps d'exécution : {stop-start} secondes")


print('Total doublons :', len(df_1_2[(df_1_2['has_similarity'] == 1)]))
print('À éliminer :', len(df_1_2[(df_1_2['has_similarity'] == 1) & ~(df_1_2['similar_with'].isna())  ]))
print('À garder :', len(df_1_2[(df_1_2['has_similarity'] == 1) & (df_1_2['similar_with'].isna())  ]))
print('Reste : ', len(df_1_2)-len(df_1_2[(df_1_2['has_similarity'] == 1) & ~(df_1_2['similar_with'].isna())  ]))

# df 1 sur 2 avec uniquement les doublons
# --------------------------------------------------------
total_doublons_1_2 = df_1_2[(df_1_2['has_similarity'] == 1)].sort_values(by='similarity_group')
print(len(total_doublons_1_2))
total_doublons_1_2[100:150]

# df 1 sur 2 sans doublons
# ------------------------------------
df_1_dd = df_1_2[df_1_2['similar_with'].isna()]
print(len(df_1_dd))
df_1_dd.sort_values(by='id_origin').head(3)

**2 sur 2** : # DEUXIEME MOITIÉ

# seq Matcher 2 sur 2 : threshold = 0.85
start = time.time()
flag_similar_posts(df_2_2)
df.head(3)
stop = time.time()
print(f"Temps d'exécution : {stop-start} secondes")

print('Total doublons :', len(df_2_2[(df_2_2['has_similarity'] == 1)]))
print('À éliminer :', len(df_2_2[(df_2_2['has_similarity'] == 1) & ~(df_2_2['similar_with'].isna())  ]))
print('À garder :', len(df_2_2[(df_2_2['has_similarity'] == 1) & (df_2_2['similar_with'].isna())  ]))
print('Reste : ', len(df_2_2)-len(df_2_2[(df_2_2['has_similarity'] == 1) & ~(df_2_2['similar_with'].isna())  ]))

# df 2 sur 2 avec uniquement les doublons
# --------------------------------------------------------
total_doublons_2_2 = df_2_2[(df_1_2['has_similarity'] == 1)].sort_values(by='similarity_group')
print(len(total_doublons_2_2))
total_doublons_2_2.head(3)

# df 2 sur 2 sans doublons
# ------------------------------------
df_2_dd = df_2_2[df_2_2['similar_with'].isna()]
print(len(df_2_dd))
df_2_dd.sort_values(by='id_origin').head(3)

# Quelques vérifications (dans cette partie, il y a deux moitiés du df)
print(len(df_1_dd))
print(len(df_2_dd))
print(len(df_2_dd)+len(df_1_dd))
#df_1.head(3)
df_2_dd.isna().sum()

---
### 2. Traitement spécifique de la partie centrale et reconcaténation de l'ensemble
---

Afin de s'assurer d'un traitement optimal des données, on procère à la même opération mais cette fois ci sur la partie centrale du jeu de données complet.

# D'abord réunir les deux moitiés
# ---------------------------------------
dfWhole=pd.concat([df_1_dd, df_2_dd])
dfWhole=dfWhole.reset_index(drop=True)
print(len(dfWhole))

display(dfWhole['text'].duplicated().sum())
print(76178/3)

# Créer 3 tiers (attn, tout doit rester trié selon sign_count)
#-----------------------------------------------
df_1tiers = dfWhole[0:25392].reset_index(drop=True)
df_CENTER = dfWhole[25392:50784].reset_index(drop=True)
df_3tiers = dfWhole[50784:].reset_index(drop=True)
print(len(df_1tiers))
print(len(df_CENTER))
print(len(df_3tiers))

# Avant SequenceMatcher, remettre les colonnes à zéro
# ------------------------------------------------
df_CENTER.loc[:,'has_similarity'] = 0
df_CENTER.loc[:,'similar_with'] = None
df_CENTER.loc[:,'similarity_group'] = None
df_CENTER.loc[:,'similarity_score'] = 0.0

df_CENTER.head(3)

# SequenceMatcher uniquement pour CENTER (threshold = 0.9)
# --------------------------------------------------------
start = time.time()
flag_similar_posts(df_CENTER)
#df_exp3.head(3)
stop = time.time()
print(f"Temps d'exécution : {stop-start} secondes")


print('Total doublons :', len(df_CENTER[(df_CENTER['has_similarity'] == 1)]))
print('À éliminer :', len(df_CENTER[(df_CENTER['has_similarity'] == 1) & ~(df_CENTER['similar_with'].isna())  ]))
print('À garder :', len(df_CENTER[(df_CENTER['has_similarity'] == 1) & (df_CENTER['similar_with'].isna())  ]))
print('Reste : ', len(df_CENTER)-len(df_CENTER[(df_CENTER['has_similarity'] == 1) & ~(df_CENTER['similar_with'].isna())  ]))

# df center avec uniquement les doublons
# --------------------------------------------------------
total_doublons_CENTER = df_CENTER[(df_CENTER['has_similarity'] == 1)].sort_values(by='similarity_group')
print(len(total_doublons_CENTER))
total_doublons_CENTER.head(3)

# df center sans doublons
# ------------------------------------
df_CENTER_dd = df_CENTER[df_CENTER['similar_with'].isna()]
print(len(df_CENTER_dd))
df_CENTER_dd.sort_values(by='id_origin').head(3)

# Concaténation finale
# ----------------------------------------------

df_Final_dd = pd.concat([df_1tiers, df_CENTER, df_3tiers], axis=0).reset_index(drop=True)
#dFFinal.to_excel('dFFinal.xlsx', index=False)
print(len(df_Final_dd))
df_Final_dd.head(3)

# Vérification et enregistrement
" ----------------------------------------------------------------------------------"
df_Final_dd.duplicated().sum()

# Synthèse
# -----------------------------------------------
print("Longueur df origine total : 102661")

print("Longueur df dédoublonné :", len(df_Final_dd))

print("Réduction totale de :", np.round((102661-len(df_Final_dd))/102661*100, 2), " %")

# A présent, on ne conserve que les posts dont la taille est située entre 50 et 3000 caractères.
filtered_df = df_Final_dd[(df_Final_dd['text'].str.len() >= 50) & (df_Final_dd['text'].str.len() <= 3000)]

print("Longueur df filtré :", len(filtered_df))

# Réinitialisation de l'index du DataFrame
filtered_df.reset_index(drop=True, inplace=True)

# Utilisation de .loc pour créer une nouvelle colonne 'id_50_3000' avec les valeurs de l'index
filtered_df.loc[:, 'id_50_3000'] = filtered_df.index

# Réorganisation des colonnes pour mettre 'id_50_3000' en premier si nécessaire
filtered_df = filtered_df[['id_50_3000', 'id_origin', 'text', 'label', 'sign_count', 'text_prepr', 'has_similarity', 'similar_with', 'similarity_group', 'similarity_score']]

filtered_df.head()

# Enregistrement en CSV
filtered_df.to_csv("../data/processed/processed_data.csv", index=False)

# Enregistrement en Excel
filtered_df.to_excel("../data/processed/processed_data.xlsx", index=False)
