==> notebooks/3.1- MAIN_Classific_supervise_SBERT.ipynb <==
# Classification supervisé - labellisation à la main - SBERT
Schéma pour 4 casses et pour 10 classes
1. (Dédoublonner)
2. Nettoyer (déjà fait)
3. Vectoriser SBERT
5. train test split
6. Training et predict (4 classes)
7. Training et predict  (10 classes)
8. Évaluations

Ce même exercice a été  mené avec TD-IDF. Ce notebook présent uniquement la vectorisation SBERT

import pandas as pd

import numpy as np


dfbrut = pd.read_excel(r'D:\Documents\DATA_SCIENCE\DATASCIENTEST_juin_2024\PROJET\SAMPLE_838_codage_XL.xlsx', index_col=0)

#dfbrut = dfbrut.drop('index', axis = 1)


dfbrut=dfbrut.reset_index()

dfbrut.columns

# 1. Vectorisation SBERT

# ------------------------------------------------------
# 3. CORPUS REUNI EN LISTE EXTERIEUR : chaque article devient un 'document' dans une liste all_docs
# --------------------------------------------------------
all_docs=  [doc for doc in dfbrut['text_prepr']]


from sentence_transformers import SentenceTransformer

model = SentenceTransformer('all-MiniLM-L6-v2')

# Step 2: Encode the documents to obtain embeddings for each multi-sentence document
document_embeddings = model.encode(all_docs)

print(document_embeddings.shape)
type(document_embeddings)

# transformer en liste de 838 vécteurs
List_doc_embeddings = document_embeddings.tolist()

len(List_doc_embeddings)

dfbrut.loc[:,'BERT_embedding']=List_doc_embeddings

dfbrut.columns

# 2. Préparation

y = dfbrut['Topic_gran_3']
X = List_doc_embeddings

dfbrut.columns

#dfbrut.to_excel('D:\Documents\DATA_SCIENCE\DATASCIENTEST_juin_2024\PROJET\SAMPLE_838_codage_XL.xlsx', index = False)

# supprimer des colonnes inutiles (issu du dédoublonnnemnt)

df = dfbrut.drop(['has_similarity', 'similar_with','similarity_group', 'similarity_score'], axis =1)



df.head(3)

# 3. Apprentissage supervisé

## a. 4 Classes (3+1) : GridSearch

from sklearn.model_selection import train_test_split, KFold, cross_validate
X_train4, X_test4, y_train4, y_test4 = train_test_split(List_doc_embeddings, y, random_state = 123, test_size = 0.2)
y_train4.value_counts()

from sklearn.model_selection import GridSearchCV
from sklearn.svm import LinearSVC
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier

clf1 = KNeighborsClassifier(n_neighbors=3)
clf2 = RandomForestClassifier(random_state=123)
clf3 = LogisticRegression(max_iter=1000)
clf4 = LinearSVC(max_iter=10000, penalty = 'l2')

param_grid_knn = {'n_neighbors': [5, 7, 9, 11]}
param_grid_rf = {'n_estimators': [20, 100, 200]}
param_grid_lr = {'C': [0.1, 1, 10, 100]}
param_grid_svc = {'C': [1, 5, 10, 15]}

grid_search_knn = GridSearchCV(clf1, param_grid_knn, cv=5)
grid_search_rf = GridSearchCV(clf2, param_grid_rf, cv=5)
grid_search_lr = GridSearchCV(clf3, param_grid_lr, cv=5)
grid_search_svc = GridSearchCV(clf4, param_grid_svc, cv=5)

grid_search_knn.fit(X_train4, y_train4)
grid_search_rf.fit(X_train4, y_train4)
grid_search_lr.fit(X_train4, y_train4)
grid_search_svc.fit(X_train4, y_train4)

# Meilleurs scores 4 classes
print("Meilleur score KNN:", grid_search_knn.best_score_)
print("Meilleur score Random Forest :", grid_search_rf.best_score_)
print("Meilleur score Régression Logistique:", grid_search_lr.best_score_)
print("Meilleur score LinearSVC Score:", grid_search_svc.best_score_)
print('----------------------')

# Meilleurs paramètres 4 classes
print("Meilleurs param. KNN :", grid_search_knn.best_params_)
print("Meilleurs param. Random Forest:", grid_search_rf.best_params_)
print("Meilleurs param. Régression Logistique:", grid_search_lr.best_params_)
print("Meilleurs param. LinearSVC :", grid_search_svc.best_params_)

# Meilleur modèle avec meilleur paramètre

lr_best = LogisticRegression(C= 1)


lr_best.fit(X_train4,y_train4)
y_pred_best_lr4=lr_best.predict(X_test4)
MXconf_lr = pd.crosstab(y_test4,y_pred_best_lr4, rownames = ['test set'], colnames = ['predicted'])
print('Accuracy LogReg : ',lr_best.score(X_test4, y_test4))
MXconf_lr


## d. Gridsearch 10 classes TF IDF :

y10 = dfbrut['Topic_gran_10']
X = List_doc_embeddings

X_train10, X_test10, y_train10, y_test10 = train_test_split(List_doc_embeddings, y10, random_state = 123, test_size = 0.2)
y_train10.value_counts()


from sklearn.svm import LinearSVC
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier

clf1 = KNeighborsClassifier(n_neighbors=3)
clf2 = RandomForestClassifier(random_state=123)
clf3 = LogisticRegression(max_iter=1000)
clf4 = LinearSVC(max_iter=10000, penalty = 'l2')

param_grid_knn = {'n_neighbors': [5, 7, 9, 11]}
param_grid_rf = {'n_estimators': [20, 100, 200]}
param_grid_lr = {'C': [0.1, 1, 10, 100]}
param_grid_svc = {'C': [1, 5, 10, 15]}

grid_search_knn = GridSearchCV(clf1, param_grid_knn, cv=5)
grid_search_rf = GridSearchCV(clf2, param_grid_rf, cv=5)
grid_search_lr = GridSearchCV(clf3, param_grid_lr, cv=5)
grid_search_svc = GridSearchCV(clf4, param_grid_svc, cv=5)

grid_search_knn.fit(X_train10, y_train10)
grid_search_rf.fit(X_train10, y_train10)
grid_search_lr.fit(X_train10, y_train10)
grid_search_svc.fit(X_train10, y_train10)

## Meilleurs scores 10 classes
print("Meilleur score KNN:", grid_search_knn.best_score_)
print("Meilleur score Random Forest :", grid_search_rf.best_score_)
print("Meilleur score Régression Logistique:", grid_search_lr.best_score_)
print("Meilleur score LinearSVC Score:", grid_search_svc.best_score_)
print('----------------------')

# Meilleurs paramètres 10 classes
print("Meilleurs param. KNN :", grid_search_knn.best_params_)
print("Meilleurs param. Random Forest:", grid_search_rf.best_params_)
print("Meilleurs param. Régression Logistique:", grid_search_lr.best_params_)
print("Meilleurs param. LinearSVC :", grid_search_svc.best_params_)

#Meilleur modèle 10 classes

svc_best = LinearSVC(max_iter=10000, penalty = 'l2', C=1)

svc_best.fit(X_train10,y_train10)
y_pred_best_svc=svc_best.predict(X_test10)
MXconf_lr = pd.crosstab(y_test10,y_pred_best_svc, rownames = ['test set'], colnames = ['predicted'])
print('Accuracy SVC linear : ',svc_best.score(X_test10, y_test10))
MXconf_lr


## évaluations 

from sklearn.metrics import davies_bouldin_score, calinski_harabasz_score

# 4 CLASSES

# -> lower is better
dbi4 = davies_bouldin_score(X_train4, y_train4)
print("Davies-Bouldin Index (DBI) 4 classes:", np.round(dbi4, 3))



# -> larger = better
ch_index4 = calinski_harabasz_score(X_train4, y_train4)
print("Calinski-Harabasz Index 4 classes:", np.round(ch_index4, 3))
print('-----------------------------------------')



# 10 CLASSES

# -> lower is better
dbi10 = davies_bouldin_score(X_train10, y_train10)
print("Davies-Bouldin Index (DBI) 10 classes:", np.round(dbi10, 3))



# -> larger = better
ch_index10 = calinski_harabasz_score(X_train10, y_train10)
print("Calinski-Harabasz Index 10 classes:", np.round(ch_index10, 3))
