==> notebooks/3.2-bertopic.ipynb <==
---------------------------------------------------------------
## VI. UTILISATION DE BERTOPIC POUR ATTRIBUER UN THÈME À CHAQUE LIGNE
--------------------------------------------------------------

Afin de pouvoir obtenir des clusters cohérents d'un point de vue thématique, nous utilisons la biliothèque Bertopics comme première méthode de classification non supervisée.

import pandas as pd

from nltk.tokenize import sent_tokenize, word_tokenize
from sentence_transformers import SentenceTransformer

from umap import UMAP
from hdbscan import HDBSCAN
from sentence_transformers import SentenceTransformer
from sklearn.feature_extraction.text import CountVectorizer

from bertopic import BERTopic
from bertopic.representation import KeyBERTInspired
from bertopic.vectorizers import ClassTfidfTransformer

from scipy.cluster import hierarchy as sch

# Import du jeu de données
df = pd.read_csv('../data/processed/processed_data.csv')

---
### 1. Préparation de Bertopic
---

# Récupération de l'intégralité des textes
texts = df['text']

# Configuration de la tokénisation
sentences = [sent_tokenize(text) for text in texts]
sentences = [sentence for doc in sentences for sentence in doc]

---
### 2. Utilisation de Bertopic
---

On choisi d'utiliser le modèle "all-mpnet-base-v2" qui présente de bon résultat sur la langue anglaise : https://huggingface.co/sentence-transformers/all-mpnet-base-v2

# Étape 1 - Extraction des embeddings
embedding_model = SentenceTransformer("all-mpnet-base-v2")
embeddings = embedding_model.encode(texts, show_progress_bar=True)

On ne démontrera pas ici l'intégralité des modalités utilisées dans le cadre du projet, une seule itération est présentée avec les paramètres communiqués ci-après.

# On défini ici les paramètres de l'UMAP et du HDBSCAN
params = [350,350,0.0,175]

# Étape 2 - Réduction de dimensionnalité
umap_model = UMAP(n_neighbors=params[0], n_components=params[1], min_dist=params[2], metric='cosine')

# Étape 3 - Établissement des clusters réduit par les empbeddings
hdbscan_model = HDBSCAN(min_cluster_size=params[3], metric='euclidean', cluster_selection_method='eom', prediction_data=True)

# Étape 4 - TOkenisation des sujet
vectorizer_model = CountVectorizer(stop_words="english")

# Étape 5 - Création des représentations des sujets
ctfidf_model = ClassTfidfTransformer()

# Étape 6 - (Optionnelle) Fine-tuning des représentations des sujets avec le modèl`bertopic.representation`
representation_model = KeyBERTInspired()

# All steps together
topic_model = BERTopic(
  embedding_model=embedding_model,          
  umap_model=umap_model,                    
  hdbscan_model=hdbscan_model,              
  vectorizer_model=vectorizer_model,        
  ctfidf_model=ctfidf_model,                
  representation_model=representation_model 
)

posts = df['text'].values

topics, probs = topic_model.fit_transform(posts)

# On visualise les sujets
topic_model.visualize_topics()

# Établissement de la Classification hiérarchique ascendante
linkage_function = lambda x: sch.linkage(x, 'single', optimal_ordering=True)
hierarchical_topics = topic_model.hierarchical_topics(texts, linkage_function=linkage_function)

topic_model.visualize_hierarchy(hierarchical_topics=hierarchical_topics)

# On extrait les informations des documents puis on les affiche
extract = topic_model.get_document_info(posts)
extract

# On enregistre le jeu de données avec ses paramètres pour une utilisation ultérieure
paramss = str(params[0]) + '_' + str(params[1]) + '_' + str(params[2]) + '_' + str(params[3])

# Enregistrement en CSV
extract.to_csv(f'../data/processed/bertopics{paramss}.csv', index=False)

# Enregistrement en Excel
extract.to_excel(f'../data/processed/bertopics{paramss}.xlsx', index=False)
